server: move service to dedicated module
