+++ artifex-server/src/lib.rs
@@ -1,0 +1,7 @@
+//
+// Copyright (C) 2023 Eric Le Bihan <eric.le.bihan.dev@free.fr>
+//
+// SPDX-License-Identifier: MIT
+//
+
+pub mod service;
